Initial support for authentication using Google Service Accounts
