+++ lib/googleAuth.js
@@ -1,52 +1,18 @@
-var fs = require("fs").promises;
-var os = require("os");
-var path = require("path");
-
-var { google } = require("googleapis");
-
-var clientID = process.env.GOOGLE_OAUTH_CLIENT_ID;
-var secret = process.env.GOOGLE_OAUTH_CONSUMER_SECRET;
-
-var tokenPath = path.join(os.homedir(), ".google_oauth_token");
-
-var getClient = async function(redirect = "http://localhost:8000/authenticate/") {
-  var auth = new google.auth.OAuth2(clientID, secret, redirect);
-  try {
-    var tokens = await loadTokenFile();
-    auth.setCredentials(tokens);
-  } catch (err) {
-    // token file won't exist on first login
+var getAuthType = function() {
+  if (process.env.GOOGLE_APPLICATION_CREDENTIALS) {
+    return "service_account";
+  } else {
+    return "oauth";
   }
+}
 
-  auth.on("tokens", updateTokenFile);
-
-  return auth;
-};
-
-var scopes = ["https://www.googleapis.com/auth/drive", "https://www.googleapis.com/auth/spreadsheets"];
-
-var loadTokenFile = async function() {
-  var tokens = await fs.readFile(tokenPath, "utf-8");
-  tokens = JSON.parse(tokens);
-  if (!tokens.access_token && !tokens.refresh_token) throw "Token file contains no actual tokens";
-  return tokens;
-};
-
-var updateTokenFile = async function(update) {
-  try {
-    var tokens = await loadTokenFile();
-  } catch (err) {
-    var tokens = {};
+var getAuthModule = function() {
+  switch (getAuthType()) {
+    case "service_account":
+      return require("./googleServiceAuth");
+    default:
+      return require("./googleOAuth");
   }
-  Object.assign(tokens, update);
-  if (!tokens.refresh_token) {
-    console.log("WARNING: No refresh_token in your Google OAuth credentials. You may have trouble staying signed in.");
-    console.log(
-      "If problems persist, visit https://myaccount.google.com/permissions and revoke permissions for this app, then reauthorize locally."
-    );
-  }
-  await fs.writeFile(tokenPath, JSON.stringify(tokens, null, 2));
-  return tokens;
-};
+}
 
-module.exports = { getClient, scopes, tokenPath, loadTokenFile, updateTokenFile };
+module.exports = { getAuthType, getAuthModule }
