+++ lib/googleServiceAuth.js
@@ -1,0 +1,13 @@
+var { google } = require("googleapis");
+
+var scopes = ["https://www.googleapis.com/auth/drive", "https://www.googleapis.com/auth/spreadsheets"];
+
+var getClient = async function() {
+  const auth = await google.auth.getClient({
+    scopes: scopes
+  });
+
+  return auth;
+};
+
+module.exports = { getClient, scopes }
