+++ lib/sheetOps.js
@@ -1,7 +1,7 @@
 var { google } = require("googleapis");
 var { promisify } = require("util");
 
-var { getClient } = require("./googleAuth");
+var { getClient } = require("./googleAuth").getAuthModule();
 
 var camelCase = function(str) {
   return str.replace(/[^\w]+(\w)/g, function(all, match) {
